// src/shared/api/generated/types/ErrorDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const errorDtoErrorMessageEnum = {
  AUTH_FAIL: "AUTH_FAIL",
  INVALID_QUERY_STRING: "INVALID_QUERY_STRING",
  ENTITY_NOT_FOUND: "ENTITY_NOT_FOUND",
  ENTITY_CREATION_FAIL: "ENTITY_CREATION_FAIL",
  NOT_ENOUGH_RIGHTS: "NOT_ENOUGH_RIGHTS",
  NOT_AUTH: "NOT_AUTH",
  EMAIL_ALREADY_IN_USE: "EMAIL_ALREADY_IN_USE",
  SLUG_ALREADY_IN_USE: "SLUG_ALREADY_IN_USE",
} as const;

export type ErrorDtoErrorMessageEnumKey =
  (typeof errorDtoErrorMessageEnum)[keyof typeof errorDtoErrorMessageEnum];

export type ErrorDto = {
  /**
   * @type string | undefined
   */
  errorMessage?: ErrorDtoErrorMessageEnumKey;
  /**
   * @type array | undefined
   */
  message?: string[];
  /**
   * @type string
   */
  error: string;
  /**
   * @type number
   */
  statusCode: number;
  /**
   * @type string | undefined
   */
  description?: string;
};